cpp: Move DeviceAllocationType to type_fwd.h
