+++ cpp/src/arrow/array/array_base.h
@@ -26,7 +26,6 @@
 #include "arrow/array/data.h"
 #include "arrow/buffer.h"
 #include "arrow/compare.h"
-#include "arrow/device_allocation_type.h"
 #include "arrow/result.h"
 #include "arrow/status.h"
 #include "arrow/type.h"
