+++ cpp/src/arrow/array/data.cc
@@ -28,7 +28,6 @@
 #include "arrow/array/util.h"
 #include "arrow/buffer.h"
 #include "arrow/device.h"
-#include "arrow/device_allocation_type.h"
 #include "arrow/scalar.h"
 #include "arrow/status.h"
 #include "arrow/type.h"
