+++ cpp/src/arrow/c/bridge.h
@@ -23,7 +23,6 @@
 
 #include "arrow/c/abi.h"
 #include "arrow/device.h"
-#include "arrow/device_allocation_type.h"
 #include "arrow/result.h"
 #include "arrow/status.h"
 #include "arrow/type_fwd.h"
