+++ cpp/src/arrow/pretty_print.cc
@@ -32,7 +32,6 @@
 
 #include "arrow/array.h"
 #include "arrow/chunked_array.h"
-#include "arrow/device_allocation_type.h"
 #include "arrow/record_batch.h"
 #include "arrow/status.h"
 #include "arrow/table.h"
